UniversityCafeteriaMenu/requirements.txt
-i https://pypi.org/simple
anyio==3.7.0 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
fastapi==0.99.0
h11==0.14.0 ; python_version >= '3.7'
httptools==0.5.0
idna==3.4 ; python_version >= '3.5'
psycopg2==2.9.6
pydantic==1.10.10 ; python_version >= '3.7'
python-dotenv==1.0.0
pyyaml==6.0
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==2.0.17
sqlalchemy-utils==0.41.1
starlette==0.27.0 ; python_version >= '3.7'
typing-extensions==4.7.0 ; python_version >= '3.7'
uvicorn[standard]==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
